+++ src/orderbook.js
@@ -1,5 +1,5 @@
 const Debug = require('debug');
-const { jsonApi } = require('./api-utils');
+const { jsonApi, wrapPaginableRequest } = require('./api-utils');
 const {
   addressSchema,
   chainIdSchema,
@@ -15,63 +15,53 @@ const debug = Debug('iexec:orderbook');
 const fetchAppOrderbook = async (
   contracts = throwIfMissing(),
   iexecGatewayURL = throwIfMissing(),
-  appAddress = throwIfMissing(),
+  app = throwIfMissing(),
   {
-    minVolume, skip, dataset, workerpool, requester, minTag, maxTag,
+    dataset, workerpool, requester, minTag, maxTag, minVolume,
   } = {},
 ) => {
-  debug({
-    minVolume,
-    skip,
-    dataset,
-    workerpool,
-    requester,
-    minTag,
-    maxTag,
-  });
   try {
-    const query = Object.assign(
-      {},
-      { chainId: await chainIdSchema().validate(contracts.chainId) },
-      {
-        app: await addressSchema({
-          ethProvider: contracts.provider,
-        }).validate(appAddress),
-      },
-      dataset && {
+    const query = {
+      chainId: await chainIdSchema().validate(contracts.chainId),
+      app: await addressSchema({
+        ethProvider: contracts.provider,
+      }).validate(app),
+      ...(dataset && {
         dataset: await addressSchema({
           ethProvider: contracts.provider,
         }).validate(dataset),
-      },
-      workerpool && {
+      }),
+      ...(workerpool && {
         workerpool: await addressSchema({
           ethProvider: contracts.provider,
         }).validate(workerpool),
-      },
-      requester && {
+      }),
+      ...(requester && {
         requester: await addressSchema({
           ethProvider: contracts.provider,
         }).validate(requester),
-      },
-      minVolume && {
+      }),
+      ...(minVolume && {
         minVolume: await positiveStrictIntSchema().validate(minVolume),
-      },
-      minTag && {
+      }),
+      ...(minTag !== undefined && {
         minTag: await tagSchema().validate(minTag),
-      },
-      maxTag && {
+      }),
+      ...(maxTag !== undefined && {
         maxTag: await tagSchema().validate(maxTag),
-      },
-      skip && {
-        skip: await positiveIntSchema().validate(skip),
-      },
-    );
-    const response = await jsonApi.get({
+      }),
+    };
+    const response = await wrapPaginableRequest(jsonApi.get)({
       api: iexecGatewayURL,
-      endpoint: '/orderbook/app',
+      endpoint: '/apporders',
       query,
     });
-    if (response.ok) return { count: response.count, appOrders: response.appOrderbook };
+    if (response.ok) {
+      return {
+        ...response,
+        appOrders: response.orders, // deprecated
+      };
+    }
     throw Error('An error occured while getting orderbook');
   } catch (error) {
     debug('fetchAppOrderbook()', error);
@@ -82,57 +72,51 @@ const fetchAppOrderbook = async (
 const fetchDatasetOrderbook = async (
   contracts = throwIfMissing(),
   iexecGatewayURL = throwIfMissing(),
-  datasetAddress = throwIfMissing(),
+  dataset = throwIfMissing(),
   {
-    minVolume, skip, app, workerpool, requester, minTag, maxTag,
+    app, workerpool, requester, minTag, maxTag, minVolume,
   } = {},
 ) => {
   try {
-    const query = Object.assign(
-      {},
-      { chainId: await chainIdSchema().validate(contracts.chainId) },
-      {
-        dataset: await addressSchema({
-          ethProvider: contracts.provider,
-        }).validate(datasetAddress),
-      },
-      app && {
+    const query = {
+      chainId: await chainIdSchema().validate(contracts.chainId),
+      dataset: await addressSchema({
+        ethProvider: contracts.provider,
+      }).validate(dataset),
+      ...(app && {
         app: await addressSchema({
           ethProvider: contracts.provider,
         }).validate(app),
-      },
-      workerpool && {
+      }),
+      ...(workerpool && {
         workerpool: await addressSchema({
           ethProvider: contracts.provider,
         }).validate(workerpool),
-      },
-      requester && {
+      }),
+      ...(requester && {
         requester: await addressSchema({
           ethProvider: contracts.provider,
         }).validate(requester),
-      },
-      minVolume && {
+      }),
+      ...(minVolume && {
         minVolume: await positiveStrictIntSchema().validate(minVolume),
-      },
-      minTag && {
+      }),
+      ...(minTag !== undefined && {
         minTag: await tagSchema().validate(minTag),
-      },
-      maxTag && {
+      }),
+      ...(maxTag !== undefined && {
         maxTag: await tagSchema().validate(maxTag),
-      },
-      skip && {
-        skip: await positiveIntSchema().validate(skip),
-      },
-    );
-    const response = await jsonApi.get({
+      }),
+    };
+    const response = await wrapPaginableRequest(jsonApi.get)({
       api: iexecGatewayURL,
-      endpoint: '/orderbook/dataset',
+      endpoint: '/datasetorders',
       query,
     });
     if (response.ok) {
       return {
-        count: response.count,
-        datasetOrders: response.datasetOrderbook,
+        ...response,
+        datasetOrders: response.orders, // deprecated
       };
     }
     throw Error('An error occured while getting orderbook');
@@ -145,49 +129,64 @@ const fetchDatasetOrderbook = async (
 const fetchWorkerpoolOrderbook = async (
   contracts = throwIfMissing(),
   iexecGatewayURL = throwIfMissing(),
-  category = throwIfMissing(),
   {
-    workerpoolAddress, minTag, signerAddress, minTrust, minVolume, skip,
+    category,
+    workerpool,
+    minTag,
+    maxTag,
+    workerpoolOwner,
+    minTrust,
+    minVolume,
+    app,
+    dataset,
   } = {},
 ) => {
   try {
-    const query = Object.assign(
-      {},
-      { chainId: await chainIdSchema().validate(contracts.chainId) },
-      { category: await uint256Schema().validate(category) },
-      workerpoolAddress && {
+    const query = {
+      chainId: await chainIdSchema().validate(contracts.chainId),
+      category: await uint256Schema().validate(category),
+      ...(workerpool && {
         workerpool: await addressSchema({
           ethProvider: contracts.provider,
-        }).validate(workerpoolAddress),
-      },
-      minTag && {
-        minTag: await tagSchema().validate(minTag),
-      },
-      signerAddress && {
-        signer: await addressSchema({
+        }).validate(workerpool),
+      }),
+      ...(app && {
+        app: await addressSchema({
           ethProvider: contracts.provider,
-        }).validate(signerAddress),
-      },
-      minTrust && {
+        }).validate(app),
+      }),
+      ...(dataset && {
+        dataset: await addressSchema({
+          ethProvider: contracts.provider,
+        }).validate(dataset),
+      }),
+      ...(minTag && {
+        minTag: await tagSchema().validate(minTag),
+      }),
+      ...(maxTag && {
+        maxTag: await tagSchema().validate(maxTag),
+      }),
+      ...(workerpoolOwner && {
+        workerpoolOwner: await addressSchema({
+          ethProvider: contracts.provider,
+        }).validate(workerpoolOwner),
+      }),
+      ...(minTrust && {
         minTrust: await positiveIntSchema().validate(minTrust),
-      },
-      minVolume && {
+      }),
+      ...(minVolume && {
         minVolume: await positiveStrictIntSchema().validate(minVolume),
-      },
-      skip && {
-        skip: await positiveIntSchema().validate(skip),
-      },
-    );
-    const response = await jsonApi.get({
+      }),
+    };
+    const response = await wrapPaginableRequest(jsonApi.get)({
       api: iexecGatewayURL,
-      endpoint: '/orderbook/workerpool',
+      endpoint: '/workerpoolorders',
       query,
     });
     if (response.ok) {
       return {
-        count: response.count,
-        openVolume: response.openVolume,
-        workerpoolOrders: response.workerpoolOrderbook,
+        ...response,
+        workerpoolOrders: response.orders, // deprecated
       };
     }
     throw Error('An error occured while getting orderbook');
@@ -200,53 +199,70 @@ const fetchWorkerpoolOrderbook = async (
 const fetchRequestOrderbook = async (
   contracts = throwIfMissing(),
   iexecGatewayURL = throwIfMissing(),
-  category = throwIfMissing(),
   {
-    requesterAddress,
-    beneficiaryAddress,
+    category,
+    requester,
+    beneficiary,
+    app,
+    dataset,
+    workerpool,
+    minTag,
     maxTag,
     maxTrust,
     minVolume,
-    skip,
   } = {},
 ) => {
   try {
-    const query = Object.assign(
-      {},
-      { chainId: await chainIdSchema().validate(contracts.chainId) },
-      { category: await uint256Schema().validate(category) },
-      requesterAddress && {
+    const query = {
+      chainId: await chainIdSchema().validate(contracts.chainId),
+      category: await uint256Schema().validate(category),
+      ...(requester && {
         requester: await addressSchema({
           ethProvider: contracts.provider,
-        }).validate(requesterAddress),
-      },
-      beneficiaryAddress && {
+        }).validate(requester),
+      }),
+      ...(beneficiary && {
         beneficiary: await addressSchema({
           ethProvider: contracts.provider,
-        }).validate(beneficiaryAddress),
-      },
-      maxTag && {
+        }).validate(beneficiary),
+      }),
+      ...(app && {
+        app: await addressSchema({
+          ethProvider: contracts.provider,
+        }).validate(app),
+      }),
+      ...(dataset && {
+        dataset: await addressSchema({
+          ethProvider: contracts.provider,
+        }).validate(dataset),
+      }),
+      ...(workerpool && {
+        workerpool: await addressSchema({
+          ethProvider: contracts.provider,
+        }).validate(workerpool),
+      }),
+      ...(minTag !== undefined && {
+        minTag: await tagSchema().validate(minTag),
+      }),
+      ...(maxTag !== undefined && {
         maxTag: await tagSchema().validate(maxTag),
-      },
-      maxTrust && {
+      }),
+      ...(maxTrust !== undefined && {
         maxTrust: await positiveIntSchema().validate(maxTrust),
-      },
-      minVolume && {
+      }),
+      ...(minVolume && {
         minVolume: await positiveStrictIntSchema().validate(minVolume),
-      },
-      skip && {
-        skip: await positiveIntSchema().validate(skip),
-      },
-    );
-    const response = await jsonApi.get({
+      }),
+    };
+    const response = await wrapPaginableRequest(jsonApi.get)({
       api: iexecGatewayURL,
-      endpoint: '/orderbook/request',
+      endpoint: '/requestorders',
       query,
     });
     if (response.ok) {
       return {
-        count: response.count,
-        requestOrders: response.requestOrderbook,
+        ...response,
+        requestOrders: response.orders, // deprecated
       };
     }
     throw Error('An error occured while getting orderbook');